repo class: make default constructor work for simple case
We need a standar way to proceed with simple creation, so take both cases
buildSelf() for existing repos(selectfrom db) and build() for non-existing ones.

Also we need default implementations that other backends can re-use or overwrite
if needed. we always have the same fields on versioncontrol_repositories table,
so it makes no sense to force all backends to re-do it.