bye bye VERSIONCONTROL_FLAG_AUTOADD_REPOSITORIES
Change interaction between repository class and backends to use inheritance
relation to acomplish that.

This commits make not mor needed _dbDeleteAdditions, _dbInsertAdditions,
_dbUpdateAdditions and _amend_repositories.

In general, repo class is modified to have insert/update/delete method that are
public final, which call a by default empty _insert/_update/_delete method that
are protected. This way, we have interactions with backends through inheritance.

The getter(VersioncontrolRepositoryCache::getRepositories) now calls a
VersioncontrolRepository::_getRepository to get special information(not in
VersioncontrolRepository:data) by repo, but maye is a good idea to make that
call by backend instead of thati for performance reasons, but that's not
possible now, so there is a TODO near the relevant code, that have exact
information about it.